tdf#56036: allow calc compiler treat tabs and newlines as spacers
Currently these characters are illegal in input formulas. This is inconsistent
with ODF spec, and brings incompatibility with other ODF implementations,
as well as with OOXML. This patch fixes this by providing corresponding
character flags for these chars (tab, cr and lf). This effectively converts
these characters to spaces. Keeping the original whitespace characters should
be done in another patch.

Change-Id: Id9362de9138ec7cd7b23c6d999181e15d5b71d5b
Reviewed-on: https://gerrit.libreoffice.org/15566
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Eike Rathke <erack@redhat.com>