dbaccess: allow storing to storage without a root one
In case of embedding a Base document, there is no root storage and
attachResource() is not called with a parameter to be able to make one,
all we can (and have to) do is to store the component to the target
store given as a parameter to storeToStorage().

With this, a .odb embedded in a .odt no longer lost on export. Import is
still broken, though.

Change-Id: I1d92a1d28f27f49609ef70fbf16effb592b9edf5