Drop misused lcl_ prefix
If used at all, lcl_ is supposed to be used for static functions truly local
to one source files. The functions here occur in several source files (well,
two, the one where they are defined and one other) so they are not "local" in
that sense. (But they could be local in the one file that uses them...) Also,
they are already in a basegfx::internal namespace.

While at it, drop the :: prefix eyesore from basegfx::internal, and align
parameter lists consistently.

Change-Id: I68b91075e0b1779ac0fa884d8f9e956f1ab7b308