don't fail on export of ooo96040-2.odt to docx
If we want to close a cell of a table which encloses another table, then close
that enclosed table first.

These are pretty pathalogical old-school tables. For now assume this can only
occur when the row of the enclosed table is cleanly closed and assert if that
isn't the case.

Change-Id: I28e4d40072cb8d92be361716e82574075b15dd89