better algorithm for OOXML column width import, tdf#91267
This patch contains a magic factor of 1.047 that is necessary as the set
value and retrieved value are not the same. When we ahve too much time
(propably never) we should explore why the value that we set for the
column width is not equal to the one that we see on the screen.

Change-Id: I317127ea5b8af2d5e8386758e66907a44fe58579