gbuild: allow recording of CppunitTests and PythonTests too
Since these don't use soffice they need to be tweaked to use RR variable.

A rr git master build from some weeks ago can record all tests, except
CppunitTest_dbaccess_firebird_test which fails.

smoketest is a bit tricky because it spawns soffice which checks RR
variable again and starts a nested rr, but fortunately there's a
flag to prevent it from aborting in this situation.

For UITests currently only the soffice.bin is recorded, not the python
test process.

The size of all the recording is about 35G per run in a --enable-dbgutil
build.

Change-Id: I2143618fa2181e36b6aaeded43637cb3481f5e47
Reviewed-on: https://gerrit.libreoffice.org/60032
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>
Tested-by: Michael Stahl <Michael.Stahl@cib.de>