lokdocview: Use maps instead of vector
Using vector each tile needs to be allocated memory irrespective of
whether tile is required or not. This approach fails when we zoom in to
a very high level to have thousands of tiles due to lot of memory
required. Using maps instead of vector takes care of this, and only
allocates Tiles when required.

Change-Id: I523f815618451a7f014e28258e0de7b1c0693370