Breeze: update icon set
Done Blue arrows - It would be nice to have a blue (or other color) used 
in wrap text, sort, sort ascending, sort descending, pivot table.

Done Spellcheck - Would be good to move the 3 green dots 1 pixel 
lower so its not touching the bottom of the ‘A’. 
(lc_SpellingAndGrammarDialog)

Done Insert Comment - All other insert buttons use a plus sign, 
so it would be good to use the same for this. In the small 
version (sc_insertannotation), there shouldnt be any plus or 
3 dots as this messes up the navigator. Would be nicer if the 
chain was a bit bigger. (lc_InsertAnnotation)

Change-Id: Ie53296c07f1c2373455a47917945ce9e0a188f76
Reviewed-on: https://gerrit.libreoffice.org/16129
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>
Tested-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>