tdf#88290 Adapt the numbering popup to more than 8 items
The items count of NumberingTypeMgr is limited by 8, but
some languages have more items, so remove that limitation.

Also fix the behavior of the 10th item. The old sidebar
popup had a "None" item, whose id was 10 (defined by
DEFAULT_NONE from include/svx/nbdtmg.hxx), and that causes
now the 10th item to behave as on/off toggle. Since we
don't have a "None" item anymore, get rid of that behavior
completely.

Change-Id: I5546ca637f9951cc093a7bbf8fda11dc35341d4e