Don't let root run unopkg without --shared option
with the exception of -h|--help.

It only makes sense to run unopkg with root priviledges if adding,
deleting etc. extensions system-wide for all users (i.e with --shared
switch ). In all other cases it only messes up access rights to
extensions, so prevent that. Furtheron, running LibO from the root
account (and installing root-user-local extensions) is not a
supported scenario.

Change-Id: I2817f62ae0bb400cefed240e792873f563fbe76e
Reviewed-on: https://gerrit.libreoffice.org/16250
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>
Tested-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>