tdf#91228: need to check the format's IsLockModified(), not the node's
commit 525c296dda460711d8798a68891706b9f82511c7 also did the extra
Remove/Add for Draw fly objects, and it turns out that that's actually
wrong because SwTextFlyCnt::SetAnchor() will set the anchor without
locking anything if it's a Draw object.  Replace it with a different
hack in SetAnchor() that applies only if it calls LockModify().

Thanks to Varun Dhall for creating a reproducer document.

Not sure if the LockModify() could be replaced completely, perhaps it's
just an optimization to avoid re-creating layout frames for the fly.

Change-Id: Ib3236f289c2c4202d48ac378a53ce02130d4ce2c