Resolves: tdf#91393 autotext (etc) not fully drawn
the paint timer is activating after we ask gtk to
size the window but before we/gtk get the ConfigureNotify
that updates gtk knowledge of the size. So the gtk drawing
calls are clipped to the previous ConfigureNotify size.

So, lets try postponing paints if we have set a size but
not received a configure notify yet

I'll also revert 850f72f94aa73480f7542af427f8326f578e050a
in favour of this attempt

Change-Id: If5e993f8e0e65053b59234fce0785398b93c1c46