lokdocview: Use GInitable
The construction of LokDocView widget can fail because it
involves initializing the lok context via lok_init.

Having lok_init calls in constructed virtual method is a bad idea
since it assumes that construction will never fail. So, implement
GInitable for this class, and move the object initialization from
constructed to initable.

Change-Id: Idf18a054cf8ef2e946392458ec52cb0107bd7454