SwViewShell::ImplEndAction: avoid direct paint
Commit f04274dd2ea73e17a9c4b76fb69fbd59e4bb4231 (sw: invalidate after
keypresses when we have a callback, 2015-01-13) only enabled
invalidate-then-paint for the tiled rendering case, because the main
loop had no priorities back then.

This limitation went away with bc9316412ca9bfd4981d467fa77d6d407e285803
(tdf#91727 - Unwind non-dispatch of idle handlers., 2015-06-10), so it's
not a problem to always do this panting inside the Paint() call.

Change-Id: I98ddfcc0f20f1553974b74ec3c859dde315612e4
Reviewed-on: https://gerrit.libreoffice.org/16587
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>