Resolves: tdf#92137 drawing layer resize undo in writer fails...
to restore location correctly

if SwFEShell::EndDrag is tweaked to not group the anchor change
undo and the draw size/pos change undo together and then you
undo each step in the ui one at a time this works fine.

The first undo occurs, then the layout kicks in on idle and calls MakeObjPos
on the drawing item and updates it for the new anchor pos

Then on triggering size undo the anchor is correct so the new
pos is set correctly.

If they are bundled together however, the layout at idle doesn't occur
until after the anchor and size have been set, so the adjustment is
too late

Change-Id: I5402d851fcae7f240cc6e9a0e15159115bb663ce