tdf#83541 Redraw the last used color after theme change
That's not what XSubToolbarController was originally designed for,
probably, but it has more in common than otherwise. A split button
offers a dropdown with some functions, and the main part of that
button updates with the last selection, to easily re-apply it.

This is based on the change made in commit
44bd22f14bbdc5fc43f260fade376ef8f8e21b68, so that updateImage
method is called once an icon refreshing is requested.

Change-Id: I6bd15638665f917ed7060886cd7835cdcd1a34f9